Added potentiometer function stub
This will later be used to collect actual potentiometer data. For now,
we're just going to use the fake value of "0.0" inside to fill the 999
protocol (key)'s exact message requirements.

Also, I'm not currently aware of what information we're trying to
collect with the potentiometer or what purpose it'll be for. So for now,
we'll just use the stub :)